+++ tests/Rewriter/SwedbankCheckDigitRewriterTest.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace byrokrat\banking\Rewriter;
+
+class SwedbankCheckDigitRewriterTest extends \PHPUnit_Framework_TestCase
+{
+    public function testRewrite()
+    {
+        $this->assertSame(
+            '1234-5,123456-7',
+            (new SwedbankCheckDigitRewriter)->rewrite('1234,5123456-7')
+        );
+    }
+}